add update conflict retry to servicebinding rotation too
