+++ internal/controller/account/servicebinding/servicebinding.go
@@ -18,13 +18,13 @@ import (
 	"github.com/sap/crossplane-provider-btp/internal"
 	servicebindingclient "github.com/sap/crossplane-provider-btp/internal/clients/account/servicebinding"
 	tfClient "github.com/sap/crossplane-provider-btp/internal/clients/tfclient"
+	"github.com/sap/crossplane-provider-btp/internal/reconcilerutil"
 	"github.com/sap/crossplane-provider-btp/internal/tracking"
 )
 
 const (
 	errNotServiceBinding    = "managed resource is not a ServiceBinding custom resource"
 	errCreateBinding        = "cannot create servicebinding"
-	errObserveSaveBinding   = "cannot save observed data"
 	errUpdateStatus         = "cannot update status"
 	errGetBinding           = "cannot get servicebinding"
 	errDeleteExpiredKeys    = "cannot delete expired keys"
@@ -149,11 +149,9 @@ func (e *external) Observe(ctx context.Context, mg resource.Managed) (managed.Ex
 	}
 
 	if observation.ResourceUpToDate && tfResource != nil {
-		if err := e.updateServiceBindingFromTfResource(cr, tfResource); err != nil {
-			return managed.ExternalObservation{}, errors.Wrap(err, errObserveSaveBinding)
-		}
-
-		if err := e.kube.Status().Update(ctx, cr); err != nil {
+		if err := reconcilerutil.UpdateStatusWithRetry(ctx, e.kube, cr, 3, func(cr *v1alpha1.ServiceBinding) {
+			_ = e.updateServiceBindingFromTfResource(cr, tfResource)
+		}); err != nil {
 			return managed.ExternalObservation{}, errors.Wrap(err, errUpdateStatus)
 		}
 	}
@@ -176,7 +174,9 @@ func (e *external) Observe(ctx context.Context, mg resource.Managed) (managed.Ex
 		return observation, nil
 	}
 
-	if err := e.kube.Status().Update(ctx, cr); err != nil {
+	if err := reconcilerutil.UpdateStatusWithRetry(ctx, e.kube, cr, 5, func(cr *v1alpha1.ServiceBinding) {
+		e.keyRotator.RetireBinding(cr)
+	}); err != nil {
 		return managed.ExternalObservation{}, errors.Wrap(err, errUpdateStatus)
 	}
 
@@ -233,11 +233,11 @@ func (e *external) Update(ctx context.Context, mg resource.Managed) (managed.Ext
 	// Clean up expired keys if there are any retired keys
 	newRetiredKeys, deleteErr := e.keyRotator.DeleteExpiredKeys(ctx, cr)
 
-	cr.Status.RetiredKeys = newRetiredKeys
-
 	// store the result in the status even if errors are returned,
 	// to remove keys for those where deletion was successfull
-	if err := e.kube.Status().Update(ctx, cr); err != nil {
+	if err := reconcilerutil.UpdateStatusWithRetry(ctx, e.kube, cr, 3, func(cr *v1alpha1.ServiceBinding) {
+		cr.Status.RetiredKeys = newRetiredKeys
+	}); err != nil {
 		return managed.ExternalUpdate{}, errors.Wrap(err, errUpdateStatus)
 	}
 	if deleteErr != nil {
